+++ jvm/src/test/scala/midvalidator/OrthoValidatorSpec.scala
@@ -14,7 +14,7 @@ class OrthoValidatorSpec extends FlatSpec {
     "MidProseABReader" ->   Vector(MidProseABDiplomatic)
   )
 
-  val repoRoot = "jvm/src/test/resources/chantsample"
+  val repoRoot = "jvm/src/test/resources/simplelatin"
   val repo = EditorsRepo(repoRoot, readerMap)
   val lib = repo.library
 
@@ -23,17 +23,26 @@ class OrthoValidatorSpec extends FlatSpec {
   //val rdr = repo.diplomaticReader(textUrn)
   //val dipl = rdr.edition(subCorpus)
 
+  val orthoMap : Map[CtsUrn, MidOrthography] = Map(
+    CtsUrn("urn:cts:chant:massordinary.eins121.text_xml:") -> Latin23,
+    CtsUrn("urn:cts:chant:massordinary.sg359.text_xml:") -> Latin23,
+    // THIS IS WRONG:
+    CtsUrn("urn:cts:chant:massordinary.sg359.neumes_xml:") -> Latin23
+  )
 
-  "An OrthographyValidator" should "require a list of MidOrthographys" in {
-    val orthoMap : Map[CtsUrn, MidOrthography] = Map(
-      CtsUrn("urn:cts:chant:massordinary.eins121.text_xml:") -> Latin23,
-      CtsUrn("urn:cts:chant:massordinary.sg359.text_xml:") -> Latin23,
-      // THIS IS WRONG:
-      CtsUrn("urn:cts:chant:massordinary.sg359.neumes_xml:") -> Latin23
-    )
-    val orthoVal = OrthographyValidator(orthoMap)
+  "An OrthographyValidator" should "find an orthography for a given URN" in {
+    val orthoVal = OrthographyValidator(lib, orthoMap)
+    assert(orthoVal.orthoForUrn(textUrn).orthography == "Simple white-space tokenization for testing")
+  }
+
+  it should "be easy to make into a TestResultGroup" in {
+    val title = "Test corpus from Latin chant project"
+    val orthoVal = OrthographyValidator(lib, orthoMap)
     val rslts = orthoVal.validate(lib)
-    println("Validation of orthography: " + rslts.size + " test results.")
+    val resultGroup = TestResultGroup(title, rslts)
+    println(resultGroup.summary)
+    import java.io.PrintWriter
+    new PrintWriter("sample-ortho-output.md"){write(resultGroup.markdown); close;}
   }
 
 }